[Feat/3] 카카오로그인을 구현합니다.

+++ app/src/main/java/com/sopt/gongbaek/data/mapper/todomain/RegisterUserInfoResponseDtoMapper.kt
@@ -4,7 +4,7 @@ import com.sopt.gongbaek.data.remote.dto.response.RegisterUserInfoResponseDto
 import com.sopt.gongbaek.domain.model.UserAuth
 
 fun RegisterUserInfoResponseDto.toDomain() = UserAuth(
-    userId = this.userId,
+    userId = this.userId.toLong(),
     accessToken = this.accessToken,
     refreshToken = this.refreshToken
 )
